# Speed up hierarchical k-means by computing distances in bulk