--- a/libs/simdvec/src/main21/java/org/elasticsearch/simdvec/internal/vectorization/PanamaESVectorUtilSupport.java
+++ b/libs/simdvec/src/main21/java/org/elasticsearch/simdvec/internal/vectorization/PanamaESVectorUtilSupport.java
@@ -822,4 +822,122 @@ public final class PanamaESVectorUtilSupport implements ESVectorUtilSupport {
         }
         return sumQuery;
     }
+
+    @Override
+    public void squareDistanceBulk(float[] query, float[] v0, float[] v1, float[] v2, float[] v3, float[] distances) {
+        FloatVector sv0 = FloatVector.zero(FLOAT_SPECIES);
+        FloatVector sv1 = FloatVector.zero(FLOAT_SPECIES);
+        FloatVector sv2 = FloatVector.zero(FLOAT_SPECIES);
+        FloatVector sv3 = FloatVector.zero(FLOAT_SPECIES);
+        final int limit = FLOAT_SPECIES.loopBound(query.length);
+        int i = 0;
+        for (; i < limit; i += FLOAT_SPECIES.length()) {
+            FloatVector qv = FloatVector.fromArray(FLOAT_SPECIES, query, i);
+            FloatVector dv0 = FloatVector.fromArray(FLOAT_SPECIES, v0, i);
+            FloatVector dv1 = FloatVector.fromArray(FLOAT_SPECIES, v1, i);
+            FloatVector dv2 = FloatVector.fromArray(FLOAT_SPECIES, v2, i);
+            FloatVector dv3 = FloatVector.fromArray(FLOAT_SPECIES, v3, i);
+            FloatVector diff0 = qv.sub(dv0);
+            sv0 = fma(diff0, diff0, sv0);
+            FloatVector diff1 = qv.sub(dv1);
+            sv1 = fma(diff1, diff1, sv1);
+            FloatVector diff2 = qv.sub(dv2);
+            sv2 = fma(diff2, diff2, sv2);
+            FloatVector diff3 = qv.sub(dv3);
+            sv3 = fma(diff3, diff3, sv3);
+        }
+        float distance0 = sv0.reduceLanes(VectorOperators.ADD);
+        float distance1 = sv1.reduceLanes(VectorOperators.ADD);
+        float distance2 = sv2.reduceLanes(VectorOperators.ADD);
+        float distance3 = sv3.reduceLanes(VectorOperators.ADD);
+
+        for (; i < query.length; i++) {
+            final float qValue = query[i];
+            final float diff0 = qValue - v0[i];
+            final float diff1 = qValue - v1[i];
+            final float diff2 = qValue - v2[i];
+            final float diff3 = qValue - v3[i];
+            distance0 = fma(diff0, diff0, distance0);
+            distance1 = fma(diff1, diff1, distance1);
+            distance2 = fma(diff2, diff2, distance2);
+            distance3 = fma(diff3, diff3, distance3);
+        }
+        distances[0] = distance0;
+        distances[1] = distance1;
+        distances[2] = distance2;
+        distances[3] = distance3;
+    }
+
+    @Override
+    public void soarDistanceBulk(
+        float[] v1,
+        float[] c0,
+        float[] c1,
+        float[] c2,
+        float[] c3,
+        float[] originalResidual,
+        float soarLambda,
+        float rnorm,
+        float[] distances
+    ) {
+
+        FloatVector projVec0 = FloatVector.zero(FLOAT_SPECIES);
+        FloatVector projVec1 = FloatVector.zero(FLOAT_SPECIES);
+        FloatVector projVec2 = FloatVector.zero(FLOAT_SPECIES);
+        FloatVector projVec3 = FloatVector.zero(FLOAT_SPECIES);
+        FloatVector acc0 = FloatVector.zero(FLOAT_SPECIES);
+        FloatVector acc1 = FloatVector.zero(FLOAT_SPECIES);
+        FloatVector acc2 = FloatVector.zero(FLOAT_SPECIES);
+        FloatVector acc3 = FloatVector.zero(FLOAT_SPECIES);
+        final int limit = FLOAT_SPECIES.loopBound(v1.length);
+        int i = 0;
+        for (; i < limit; i += FLOAT_SPECIES.length()) {
+            FloatVector v1Vec = FloatVector.fromArray(FLOAT_SPECIES, v1, i);
+            FloatVector c0Vec = FloatVector.fromArray(FLOAT_SPECIES, c0, i);
+            FloatVector c1Vec = FloatVector.fromArray(FLOAT_SPECIES, c1, i);
+            FloatVector c2Vec = FloatVector.fromArray(FLOAT_SPECIES, c2, i);
+            FloatVector c3Vec = FloatVector.fromArray(FLOAT_SPECIES, c3, i);
+            FloatVector originalResidualVec = FloatVector.fromArray(FLOAT_SPECIES, originalResidual, i);
+            FloatVector djkVec0 = v1Vec.sub(c0Vec);
+            FloatVector djkVec1 = v1Vec.sub(c1Vec);
+            FloatVector djkVec2 = v1Vec.sub(c2Vec);
+            FloatVector djkVec3 = v1Vec.sub(c3Vec);
+            projVec0 = fma(djkVec0, originalResidualVec, projVec0);
+            projVec1 = fma(djkVec1, originalResidualVec, projVec1);
+            projVec2 = fma(djkVec2, originalResidualVec, projVec2);
+            projVec3 = fma(djkVec3, originalResidualVec, projVec3);
+            acc0 = fma(djkVec0, djkVec0, acc0);
+            acc1 = fma(djkVec1, djkVec1, acc1);
+            acc2 = fma(djkVec2, djkVec2, acc2);
+            acc3 = fma(djkVec3, djkVec3, acc3);
+        }
+        float proj0 = projVec0.reduceLanes(ADD);
+        float dsq0 = acc0.reduceLanes(ADD);
+        float proj1 = projVec1.reduceLanes(ADD);
+        float dsq1 = acc1.reduceLanes(ADD);
+        float proj2 = projVec2.reduceLanes(ADD);
+        float dsq2 = acc2.reduceLanes(ADD);
+        float proj3 = projVec3.reduceLanes(ADD);
+        float dsq3 = acc3.reduceLanes(ADD);
+        // tail
+        for (; i < v1.length; i++) {
+            float v = v1[i];
+            float djk0 = v - c0[i];
+            float djk1 = v - c1[i];
+            float djk2 = v - c2[i];
+            float djk3 = v - c3[i];
+            proj0 = fma(djk0, originalResidual[i], proj0);
+            proj1 = fma(djk1, originalResidual[i], proj1);
+            proj2 = fma(djk2, originalResidual[i], proj2);
+            proj3 = fma(djk3, originalResidual[i], proj3);
+            dsq0 = fma(djk0, djk0, dsq0);
+            dsq1 = fma(djk1, djk1, dsq1);
+            dsq2 = fma(djk2, djk2, dsq2);
+            dsq3 = fma(djk3, djk3, dsq3);
+        }
+        distances[0] = dsq0 + soarLambda * proj0 * proj0 / rnorm;
+        distances[1] = dsq1 + soarLambda * proj1 * proj1 / rnorm;
+        distances[2] = dsq2 + soarLambda * proj2 * proj2 / rnorm;
+        distances[3] = dsq3 + soarLambda * proj3 * proj3 / rnorm;
+    }
 }