This commit adds on-heap bulk distance computations. In particular, it implements the methods `ESVectorUtil#squareDistanceBulk` and ``ESVectorUtil#soarDistanceBulk` to compute four distances in one method call. Microbenchmarks shows a nice speed up, for example for AVX2:

```
Benchmark                                   (dims)   Mode  Cnt  Score   Error   Units
SquareDistanceBenchmark.soarDistance           384  thrpt    5  4.387 ± 0.167  ops/ms
SquareDistanceBenchmark.soarDistance           782  thrpt    5  1.952 ± 0.357  ops/ms
SquareDistanceBenchmark.soarDistance          1024  thrpt    5  1.658 ± 0.817  ops/ms
SquareDistanceBenchmark.soarDistanceBulk       384  thrpt    5  6.627 ± 0.292  ops/ms
SquareDistanceBenchmark.soarDistanceBulk       782  thrpt    5  3.577 ± 0.255  ops/ms
SquareDistanceBenchmark.soarDistanceBulk      1024  thrpt    5  3.171 ± 0.360  ops/ms
SquareDistanceBenchmark.squareDistance         384  thrpt    5  5.853 ± 0.519  ops/ms
SquareDistanceBenchmark.squareDistance         782  thrpt    5  2.844 ± 0.034  ops/ms
SquareDistanceBenchmark.squareDistance        1024  thrpt    5  2.515 ± 0.104  ops/ms
SquareDistanceBenchmark.squareDistanceBulk     384  thrpt    5  8.669 ± 1.235  ops/ms
SquareDistanceBenchmark.squareDistanceBulk     782  thrpt    5  4.012 ± 0.449  ops/ms
SquareDistanceBenchmark.squareDistanceBulk    1024  thrpt    5  3.360 ± 0.454  ops/ms
``` 

The commit updates k-means local to use those new methods which shows a nice speed up. For example, indexing 3 million GLOVE vectors with 200 dimensions.

before:
```
index_name                         index_type  num_docs  index_time(ms)  force_merge_time(ms)  num_segments
---------------------------------  ----------  --------  --------------  --------------------  ------------  
enwiki-20120502-lines-1k-200d.vec         ivf   3000008           39798                 81817             0
```

after:
```

index_name                         index_type  num_docs  index_time(ms)  force_merge_time(ms)  num_segments
---------------------------------  ----------  --------  --------------  --------------------  ------------  
enwiki-20120502-lines-1k-200d.vec         ivf   3000008           29074                 56799             0
```

Or 1 million Cohere vectors with 1024 dimensions:

before:

```
index_name       index_type  num_docs  index_time(ms)  force_merge_time(ms)  num_segments
---------------  ----------  --------  --------------  --------------------  ------------  
wiki1024en.docs         ivf   1000008           51073                106968             0
```

after:

```
index_name       index_type  num_docs  index_time(ms)  force_merge_time(ms)  num_segments
---------------  ----------  --------  --------------  --------------------  ------------  
wiki1024en.docs         ivf   1000008           32692                 85660             0
```